Handle state space failures
If the ArgStates plugin fails to complete its analysis and crashes for a
symbol we need to exclude any assumptions about this symbol. E.g. if
analysis for the symbol 'foo' finds only constant values in one SubDirTU and
crashes for all others, we need to defensively assume that it could be
non-det in one of the TUs were analysis failed.